+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/event/process/EventProcessJob.java
@@ -160,8 +160,8 @@ public class EventProcessJob implements Job {
 			createOutputAppender();
 			initializeStreamingAppender();
 
-			float totalFiles = files.length;
-			float totalDone = 0;
+			long totalFiles = files.length;
+			long totalDone = 0;
 
 			// read each file and process lines
 			try {
@@ -187,7 +187,7 @@ public class EventProcessJob implements Job {
 					}
 					totalDone++;
 					logger.info("{}/{} files processed - {}% done", totalDone, totalFiles,
-							Math.round((totalDone / totalFiles) * 100));
+							Math.round(((float)totalDone / (float)totalFiles) * 100));
 				}
 			} finally {
 				// make sure all close are called, hence the horror below of nested finally blocks
@@ -263,7 +263,7 @@ public class EventProcessJob implements Job {
 
 		LineNumberReader lnr = new LineNumberReader(new FileReader(file));
 		lnr.skip(Long.MAX_VALUE);
-		float totalLines = lnr.getLineNumber() + 1; //Add 1 because line index starts at 0
+		long totalLines = lnr.getLineNumber() + 1; //Add 1 because line index starts at 0
 		lnr.close();
 
 		try {
@@ -284,7 +284,7 @@ public class EventProcessJob implements Job {
 					}
 					if (linesPrintEnabled && numOfLines % linesPrintSkip == 0) {
 						logger.info("{}/{} lines processed - {}% done", numOfLines, totalLines,
-								Math.round(((float)numOfLines / totalLines) * 100));
+								Math.round(((float)numOfLines / (float)totalLines) * 100));
 					}
 				}
 			}
